ai-chatbot/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
websockets==14.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
groq==0.15.0
httpx==0.28.1
ollama==0.4.4
openpyxl==3.1.5
python-docx==1.1.2
reportlab==4.2.5
pdfplumber==0.11.4
pandas==2.2.3
easyocr==1.7.2
Pillow==11.1.0
ffmpeg-python==0.2.0
faster-whisper==1.1.0
edge-tts==7.0.0
duckduckgo-search==7.2.1
requests==2.32.3
pydantic-settings==2.7.0
filetype==1.2.0
